+++ test/Option.ts
@@ -1,7 +1,8 @@
-import { unsafeUnwrap, noneAs, invert } from "../src/Option"
+import { unsafeUnwrap, noneAs, invert, toMonoid } from "../src/Option"
 import * as O from "fp-ts/Option"
 import * as S from "fp-ts/string"
 import fc from "fast-check"
+import { pipe } from "fp-ts/function"
 
 describe("Option", () => {
   describe("unsafeUnwrap", () => {
@@ -45,6 +46,18 @@ describe("Option", () => {
       fc.assert(
         fc.property(fc.string(), x => expect(f(x)(O.some(x))).toEqual(O.none)),
       )
+    })
+  })
+
+  describe("toMonoid", () => {
+    const f = toMonoid(S.Monoid)
+
+    it("returns monoidal identity on None", () => {
+      expect(f(O.none)).toEqual(S.Monoid.empty)
+    })
+
+    it("unwraps Some", () => {
+      fc.assert(fc.property(fc.string(), x => pipe(x, O.some, f) === x))
     })
   })
 })